+++ CHANGELOG.md
@@ -18,6 +18,7 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 - Updated Suggest Tags input to accept freeInput option (defaults to true) - when set to false only provided typeahead tags are valid
 - Set default title and notifications to Integral Backend CRUD endpoints
 - Pages - Added 'Archived' status
+- Lighthouse Auditing
 
 ### Fixed